Update README.md
As an Intern, I have created and completed in 10-11 days in Jupyter Lab as per the task of Skill Craft Technology i.e., DATA SCIENCE ----TASK 1. Task : Create a bar chart or histogram to visualize the distribution of a categorical or continuous variable, such as the distribution of ages or genders in a population.